Moved Libraries into their own folder.
