+++ SuperGameEngine/FatedQuest.Libraries/FileHandling/framework.h
