+++ SuperGameEngine/FatedQuest.Libraries/Logging/Logger.cpp
