+++ SuperGameEngine/FatedQuest.Libraries/Logging/Logger.h
