+++ SuperGameEngine/FatedQuest.Libraries/Logging/Logging.vcxproj.filters
