+++ SuperGameEngine/FatedQuest.Libraries/MathHelpers/MathHelpers.cpp
