+++ SuperGameEngine/FatedQuest.Libraries/MathHelpers/framework.h
