+++ SuperGameEngine/FatedQuest.Libraries/MathHelpers/pch.cpp
