+++ SuperGameEngine/FatedQuest.Libraries/MathHelpers/pch.h
