+++ SuperGameEngine/FatedQuest.Libraries/StandardCLibrary/FString.cpp
