+++ SuperGameEngine/FatedQuest.Libraries/StandardCLibrary/FString.h
